Create readme
This is a very simple VR room experience for those who are fear of heights. Basically, this is a plank walking experience, with some add-ons to make the experience more challenging. The objective is to grab a gold medal which is at the top of a high building. This is my first development for the Oculus Quest 2 VR headset. I enjoyed the development a lot. I think is impressive what is possible to do with free tools such as Blender and Unreal Engine. I want to share this app and encourage anyone interested in start creating content for the new generation of VR headsets. Hope you enjoy it. Javi.